+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] 2019-08-12
+
+### Added
+
+- Backup functionality. A config key specifies how many backup coordinate files are
+saved. If there are too many files, the oldest backup file is deleted.
+
 ## [1.1.0] 2019-08-12
 
 ### Changed